[AAE-5690] move services to separate folder
