+++ src/app/stories/login/login.component.ts
@@ -1,7 +1,7 @@
 import { AppConfigService } from '@alfresco/adf-core';
 import { Component, Input, OnInit, ViewChild } from '@angular/core';
 import { Validators } from '@angular/forms';
-import { setIsOuathReturnValue } from './authentication-stub.service';
+import { setIsOuathReturnValue } from './mock/authentication-stub.service';
 
 @Component({
   selector: 'aca-login',
@@ -9,44 +9,44 @@ import { setIsOuathReturnValue } from './authentication-stub.service';
 })
 export class LoginComponent implements OnInit {
   /**
-   * Should the Remember me checkbox be shown? When selected, this option will remember the logged-in user after the browser is closed to avoid logging in repeatedly.
-   */
+  * Should the Remember me checkbox be shown? When selected, this option will remember the logged-in user after the browser is closed to avoid logging in repeatedly.
+  */
   @Input()
   showRememberMe: boolean;
 
   /**
-   * Should the extra actions (Need Help, Register, etc) be shown?
-   */
+  * Should the extra actions (Need Help, Register, etc) be shown?
+  */
   @Input()
   showLoginActions: boolean;
 
   /**
-   * The copyright text below the login box.
-   */
+  * The copyright text below the login box.
+  */
   @Input()
   copyrightText: string;
 
   /**
-   * Path to a custom logo image.
-   */
+  * Path to a custom logo image.
+  */
   @Input()
   logoImageUrl: string;
 
   /**
-   * Path to a custom background image.
-   */
+  * Path to a custom background image.
+  */
   @Input()
   backgroundImageUrl: string;
 
   /**
-   * Sets the URL of the NEED HELP link in the footer.
-   */
+  * Sets the URL of the NEED HELP link in the footer.
+  */
   @Input()
   needHelpLink: string;
 
   /**
-   * Sets the URL of the REGISTER link in the footer.
-   */
+  * Sets the URL of the REGISTER link in the footer.
+  */
   @Input()
   registerLink: string;
 
@@ -87,8 +87,8 @@ export class LoginComponent implements OnInit {
   alfrescoLogin: any;
 
   /**
-   * Custom validation rules for the login form.
-   */
+  * Custom validation rules for the login form.
+  */
   fieldsValidation: any;
 
   constructor(private appConfigService: AppConfigService) {}
