+++ src/app/stories/login/mock/authentication-stub.service.ts
